requirements.txt
# Core app dependencies
streamlit==1.34.0
opencv-python-headless==4.9.0.80
torch==2.1.2
torchvision==0.16.2
numpy==1.24.4
Pillow==10.2.0

# Optional but useful
tqdm==4.66.1